+++ app/helpers/project_helper.rb
@@ -49,10 +49,13 @@ module ProjectHelper
     tag.address address.compact.join("<br/>").html_safe, class: %w[govuk-address]
   end
 
-  def completed_project_notification_banner(project)
+  def project_notification_banner(project, user)
     if project.completed?
       render NotificationBanner
         .new(message: t("project.notifications.completed_html", date: project.completed_at.to_formatted_s(:govuk_date_time_date_only)))
+    elsif project.assigned_to != user || project.assigned_to.nil?
+      render NotificationBanner
+        .new(message: t("project.notifications.not_assigned_html"))
     end
   end
 end