+++ spec/helpers/project_helper_spec.rb
@@ -167,21 +167,64 @@ RSpec.describe ProjectHelper, type: :helper do
     end
   end
 
-  describe "#completed_project_notification_banner" do
+  describe "#project_notification_banner" do
     context "when the project is not complete i.e. in-progress" do
-      it "does nothing" do
-        project = build(:conversion_project, completed_at: nil)
+      context "when the project is assigned to the user" do
+        it "does nothing" do
+          user = build(:user)
+          project = build(:conversion_project, completed_at: nil, assigned_to: user)
 
-        expect(completed_project_notification_banner(project)).to be_nil
+          expect(project_notification_banner(project, user)).to be_nil
+        end
+      end
+
+      context "when the project is not assigned to the user" do
+        it "renders a notification banner" do
+          user = build(:user)
+          other_user = build(:user, email: "other.user@education.gov.uk")
+
+          project = build(:conversion_project, completed_at: nil, assigned_to: other_user)
+
+          expect(project_notification_banner(project, user))
+            .to include("Not assigned to project")
+        end
+      end
+
+      context "when the project is not assigned to any user" do
+        it "renders a notification banner" do
+          user = build(:user)
+
+          project = build(:conversion_project, completed_at: nil, assigned_to: nil)
+
+          expect(project_notification_banner(project, user))
+            .to include("Not assigned to project")
+        end
       end
     end
 
     context "when the project is completed" do
-      it "renders the notification banner" do
-        project = build(:conversion_project, completed_at: Date.yesterday)
+      context "when the project is assigned to the user" do
+        it "renders the compeleted notification banner" do
+          user = build(:user)
+          other_user = build(:user, email: "other.user@education.gov.uk")
 
-        expect(completed_project_notification_banner(project))
-          .to include("Project completed")
+          project = build(:conversion_project, completed_at: Date.today - 3.months, assigned_to: other_user)
+
+          expect(project_notification_banner(project, user))
+            .to include("Project completed")
+        end
+      end
+
+      context "when the project is not assigned to the user" do
+        it "renders completed notification banner" do
+          user = build(:user)
+          other_user = build(:user, email: "other.user@education.gov.uk")
+
+          project = build(:conversion_project, completed_at: Date.today - 3.months, assigned_to: other_user)
+
+          expect(project_notification_banner(project, user))
+            .to include("Project completed")
+        end
       end
     end
   end